requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langchain-core==0.3.69
langchain-openai==0.3.28
langsmith==0.4.8
MarkupSafe==3.0.2
narwhals==1.47.1
numpy==2.0.2
openai==1.97.0
orjson==3.11.0
packaging==25.0
pandas==2.3.1
pgvector==0.4.1
pillow==11.3.0
protobuf==6.31.1
psycopg2-binary==2.9.9
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.26.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.47.0
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
zstandard==0.23.0
